added player exclusiove functionality and made some test builds
