+++ Unity/Assets/Scenes/MobileControlScene.unity
@@ -938,7 +938,7 @@ VideoPlayer:
   - {fileID: 0}
   m_DirectAudioVolumes:
   - 1
-  m_Url: http://192.168.1.245:8889/mystream/whep
+  m_Url: http://192.168.1.245:8889/stream/whep
   m_EnabledAudioTracks: 01
   m_DirectAudioMutes: 00
   m_ControlledAudioTrackCount: 1
