+++ Unity/Assets/Scenes/SpectatorScene.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 5d163b4314d4649bea1c51e7db865ffd
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
